# Optimize Attention Mechanism Performance