## Overview
This PR introduces performance optimizations to the attention mechanism, specifically targeting the `DotProductAttention` function. The changes result in a 30% improvement in execution time while maintaining the same memory footprint.

## Changes
- Pre-calculate the scaling factor (`1.0 / math.Sqrt(float64(d_k))`) outside the main loop
- Skip value accumulation for zero attention weights
- Add comprehensive dimension validation checks
- Maintain numerical stability in Softmax through max-value subtraction
- Add benchmark tests with fixed seeds for reproducible performance measurements

## Performance Impact
Benchmark results show significant improvements:

| Benchmark                  | Before (ns/op) | After (ns/op) | Change |
|---------------------------|----------------|---------------|--------|
| DotProductAttention       | 14,842         | 10,442        | -29.6% |
| Softmax                   | 9,048          | 9,056         | +0.1%  |
| DotProduct                | 1,282          | 1,285         | +0.2%  |

Memory usage (B/op and allocs/op) remains unchanged.

## Testing
- Added comprehensive benchmark tests with fixed seeds
- All existing tests pass
- Performance improvements verified through benchmark comparisons

## Notes
- No API changes were made
- Memory usage remains constant
- Code readability and maintainability preserved